requirements.txt
Django==2.2.2
djangorestframework==3.9.4
whitenoise==4.1.2
gunicorn==19.9.0
psycopg2==2.8.2
Pillow==6.0.0
dj_database_url==0.4.2
python-decouple==3.0
boto==2.40.0
django-storages==1.5.2
unicodecsv==0.14.1